requirements.txt
web3
python-dotenv
confluent-kafka